tidy-up: misc C and CMake nits
- CMake/win32-cache.cmake: sort and sync MSVC pre-fills with mingw-w64.

- curl_sasl: use `uint32_t` for local `httpauth` value. To match struct.
  Follow-up to e3691612ea1acb28e77fee6b4b2b9da0ac2c8f8b #20209

- libstests: move literal to the right-hand side of comparisons.

- windows: drop `!= 0`/`== 0` from `WSAEventSelect()` result checks,
  to keep it short.

- tool_cfgable: replace `CURL_GLOBAL_ALL` with `CURL_GLOBAL_DEFAULT` in
  debug builds to match the macro used in non-debug and libcurl init.
  (Both macros map to the same literal value; there is no code change.)

- perf/urlparser: use double-const in test array.

- examples: `CURL_GLOBAL_DEFAULT` -> `CURL_GLOBAL_ALL` to match rest of
  examples.

- libtests: `CURL_GLOBAL_DEFAULT` -> `CURL_GLOBAL_ALL` to match rest of
  tests.

Closes #22783

+++ lib/curl_sasl.c
@@ -142,7 +142,7 @@ CURLcode Curl_sasl_parse_url_auth_option(struct SASL *sasl,
 void Curl_sasl_init(struct SASL *sasl, struct Curl_easy *data,
                     const struct SASLproto *params)
 {
-  unsigned long auth = data->set.httpauth;
+  uint32_t auth = data->set.httpauth;
 
   sasl->params = params;           /* Set protocol dependent parameters */
   sasl->state = SASL_STOP;         /* Not yet running */
